Refactor to support command conflict
The main logic is separated into two parts: backend + frontend. Backend
is in charge of managing multiple package sources (repository). There
are:
- 1 dropin repository
- 1 default repository that are synchronized with remote registry json
- N repositories that can be used to synchronize with other remote
  registry.

  This review only support 1 dropin + 1 default repository

  Frontend will use the API from backend to construct the command
  launcher interface.

+++ test/integration.sh
@@ -5,6 +5,8 @@ SCRIPT_DIR=${SCRIPT_DIR:-$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/
 SCRIPT_DIR=${SCRIPT_DIR//\\//}
 echo "integration test directory: $SCRIPT_DIR"
 
+KEEP_OUTPUT=${KEEP_OUTPUT:-"no"}
+
 EXIT_CODE=0
 TEST_COUNT=0
 FAILURE_COUNT=0
@@ -91,8 +93,12 @@ fi
 ##
 # remove the output folder
 ##
-echo "clean up"
-rm -rf $OUTPUT_DIR
+if [ $KEEP_OUTPUT == "yes" ]; then
+  echo "KEEP_OUTPUT = yes, skip clean up"
+else
+  echo "clean up"
+  rm -rf $OUTPUT_DIR
+fi
 
 echo ""
 echo "Total test suits $TEST_COUNT, failure $FAILURE_COUNT"
